Conflicts:
	Examples/ExampleStapling/StaplingExample.cpp
	SurgSim/Physics/UnitTests/DeformableCollisionRepresentationTest.cpp

+++ SurgSim/Blocks/MassSpringNDRepresentationUtils.h
@@ -41,7 +41,7 @@ namespace Blocks
 /// \return The newly create spring
 std::shared_ptr<SurgSim::Physics::LinearSpring> createLinearSpring(
 	const std::shared_ptr<SurgSim::Math::OdeState> state,
-	unsigned int nodeId0, unsigned int nodeId1,
+	size_t nodeId0, size_t nodeId1,
 	double stiffness, double damping);
 
 };  // namespace Blocks
